specific folder for spikeglx and open ephys
