+++ tests/data/openephys/OE_5_Neuropix-PXI-multi-probe/.ipynb_checkpoints/settings-checkpoint.xml
