+++ tests/data/openephys/OE_5_Neuropix-PXI-multi-probe/settings.xml
