+++ tests/data/openephys/OE_Neuropix-PXI-multi-probe/settings.xml
