+++ tests/data/openephys/OE_Neuropix-PXI-multi-probe/settings_2.xml
