+++ tests/data/openephys/OE_Neuropix-PXI/settings.xml
