+++ tests/data/spikeglx/Day_3_g0_t0.imec1.ap.meta
