+++ tests/data/spikeglx/NP24_g0_t0.imec0.ap.meta
