+++ tests/data/spikeglx/NeuropixelPhase3A_file_g0_t0.imec.ap.meta
