+++ tests/data/spikeglx/Noise_g0_t0.imec0.ap.meta
