+++ tests/data/spikeglx/TEST_20210920_0_g0_t0.imec0.ap.meta
