+++ tests/data/spikeglx/doppio-checkerboard_t0.imec0.ap.meta
