+++ tests/data/spikeglx/p2_g0_t0.imec0.ap.meta
