+++ tests/test_io/test_io.py
@@ -8,8 +8,6 @@ from probeinterface import read_prb, write_prb
 from probeinterface import (
     read_imro,
     write_imro,
-    parse_spikeglx_meta,
-    get_saved_channel_indices_from_spikeglx_meta,
 )
 from probeinterface import generate_dummy_probe_group
 
@@ -201,29 +199,6 @@ def test_prb():
     # import matplotlib.pyplot as plt
     # plot_probe(probe)
     # plt.show()
-
-
-def test_parse_meta():
-    for meta_file in [
-        "doppio-checkerboard_t0.imec0.ap.meta",
-        "Day_3_g0_t0.imec1.ap.meta",
-        "allan-longcol_g0_t0.imec0.ap.meta",
-    ]:
-        meta = parse_spikeglx_meta(data_path / meta_file)
-
-
-def test_get_saved_channel_indices_from_spikeglx_meta():
-    # all channel saved + 1 synchro
-    chan_inds = get_saved_channel_indices_from_spikeglx_meta(
-        data_path / "Noise_g0_t0.imec0.ap.meta"
-    )
-    assert chan_inds.size == 385
-
-    # example by Pierre Yger NP1.0 with 384 but only 151 channels are saved + 1 synchro
-    chan_inds = get_saved_channel_indices_from_spikeglx_meta(
-        data_path / "Day_3_g0_t0.imec1.ap.meta"
-    )
-    assert chan_inds.size == 152
 
 
 def test_readimro():
