+++ tests/test_io/test_spikeglx.py
@@ -3,9 +3,13 @@ import numpy as np
 
 import pytest
 
-from probeinterface import read_spikeglx
+from probeinterface import (
+    read_spikeglx,
+    parse_spikeglx_meta,
+    get_saved_channel_indices_from_spikeglx_meta,
+)
 
-data_path = Path(__file__).absolute().parent.parent / "data"
+data_path = Path(__file__).absolute().parent.parent / "data" / "spikeglx"
 
 
 def test_NP1():
@@ -64,3 +68,26 @@ def tes_NP1_384_channels():
     assert probe.get_shank_count() == 1
     assert probe.get_contact_count() == 151
     assert 152 not in probe.contact_annotations["channel_ids"]
+
+
+def test_parse_meta():
+    for meta_file in [
+        "doppio-checkerboard_t0.imec0.ap.meta",
+        "Day_3_g0_t0.imec1.ap.meta",
+        "allan-longcol_g0_t0.imec0.ap.meta",
+    ]:
+        meta = parse_spikeglx_meta(data_path / meta_file)
+
+
+def test_get_saved_channel_indices_from_spikeglx_meta():
+    # all channel saved + 1 synchro
+    chan_inds = get_saved_channel_indices_from_spikeglx_meta(
+        data_path / "Noise_g0_t0.imec0.ap.meta"
+    )
+    assert chan_inds.size == 385
+
+    # example by Pierre Yger NP1.0 with 384 but only 151 channels are saved + 1 synchro
+    chan_inds = get_saved_channel_indices_from_spikeglx_meta(
+        data_path / "Day_3_g0_t0.imec1.ap.meta"
+    )
+    assert chan_inds.size == 152
